Changed order of e_detail and t_detail
There seems to be an issue with how we are using af::timeit(). I understand that it will do a bunch of tests and calculate the median time for the function to execute, but it's also changing the value being stored for pi_v1(). For example, had I left the order the way it was, then I would get an error of like 113 instead of the normal error of <0.001.
I'm investigating this right now, but I just wanted to mention my thoughts on this in the meantime.