Fixed block/grid_size overwrite issue
The init_v2 was called right after init_v1, so block_size and grid_size were being overwritten. This resulted in slower times for pi_v1 because it was using the block_size and grid_size intended for pi_v2.
I fixed this by making two more namespace variables: block_size2 and grid_size2. It's kind of a clunky fix at the moment, but okay for now.